organizing
Changed the display from flex to fixed, on the left and right. Made the buttons centered, and fixed the main layout of the code. Changed some of the classes and added a lot of divs to put stuff into later. Created a css folder just for the styling and separated the html from the css. Sorted out the order of the code so its easier to read.